backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo[srv]==4.6.1
gunicorn==21.2.0
python-dotenv==1.0.0
